requirements.txt
openai-whisper
pyannote.audio>=4.0.4
pyannote.core
torch
torchaudio
sounddevice
scipy
numpy
python-dotenv
reportlab
fastapi
uvicorn[standard]
python-multipart
